requirements.txt
pandas
numpy
keras
tensorflow
h5py
flask
Pillow
gunicorn